+++ trunk/auto/modules.sh
@@ -3,8 +3,7 @@
 #
 # params:
 #     $SRS_OBJS the objs directory to store the Makefile. ie. ./objs
-#     $SRS_OBJS_DIR the objs directory for Makefile. ie. objs
-#     $SRS_MAKEFILE the makefile name. ie. Makefile
+#     $SRS_OBJS the objs directory for Makefile. ie. objs
 #
 #     $MODULE_DIR the module dir. ie. src/os/linux
 #     $MODULE_ID the id of module. ie. CORE
@@ -16,7 +15,7 @@
 # returns:
 #     $MODULE_OBJS array, the objects of the modules, used for link the binary
 
-FILE=${SRS_OBJS}/${SRS_MAKEFILE}
+FILE=${SRS_OBJS}/Makefile
 echo "#####################################################################################" >> ${FILE}
 echo "# The module ${MODULE_ID}." >> ${FILE}
 echo "#####################################################################################" >> ${FILE}
@@ -27,7 +26,7 @@ echo "# INCS for ${MODULE_ID}, headers of module and its depends to compile" >> 
 #
 # the public include files, for example:
 #       CORE_MODULE_INCS = -Isrc/core
-echo "${MODULE_ID}_MODULE_INCS = -I${MODULE_DIR} " >> ${FILE}
+echo "${MODULE_ID}_MODULE_INCS = -I${SRS_WORKDIR}/${MODULE_DIR} " >> ${FILE}
 #
 # the private include files, for example:
 #       CORE_INCS = -Isrc/core -Iobjs
@@ -36,7 +35,7 @@ echo "${MODULE_ID}_MODULE_INCS = -I${MODULE_DIR} " >> ${FILE}
 INCS_NAME="${MODULE_ID}_INCS"
 #
 # current module header files
-echo -n "${INCS_NAME} = -I${MODULE_DIR} " >> ${FILE}
+echo -n "${INCS_NAME} = -I${SRS_WORKDIR}/${MODULE_DIR} " >> ${FILE}
 #
 # depends module header files
 for item in ${MODULE_DEPENDS[*]}; do
@@ -60,8 +59,7 @@ echo "# DEPS for ${MODULE_ID}, the depends of make schema" >> ${FILE}
 DEPS_NAME="${MODULE_ID}_DEPS"
 echo -n "${DEPS_NAME} = " >> ${FILE}
 for item in ${MODULE_FILES[*]}; do
-    HEADER_FILE="${MODULE_DIR}/${item}.hpp"
-    if [[ ! -f ${HEADER_FILE} ]]; then HEADER_FILE="${MODULE_DIR}/${item}.h"; fi
+    HEADER_FILE="${SRS_WORKDIR}/${MODULE_DIR}/${item}.hpp"
     if [ -f ${HEADER_FILE} ]; then
         echo -n " ${HEADER_FILE}" >> ${FILE}
     fi
@@ -78,10 +76,9 @@ echo "" >> ${FILE}; echo "" >> ${FILE}
 echo "# OBJ for ${MODULE_ID}, each object file" >> ${FILE}
 MODULE_OBJS=()
 for item in ${MODULE_FILES[*]}; do
-    CPP_FILE="${MODULE_DIR}/${item}.cpp"
-    if [[ ! -f ${CPP_FILE} ]]; then CPP_FILE="${MODULE_DIR}/${item}.cc"; fi
-    OBJ_FILE="${SRS_OBJS_DIR}/${MODULE_DIR}/${item}.o"
-    MODULE_OBJS="${MODULE_OBJS[@]} ${CPP_FILE}"
+    CPP_FILE="${SRS_WORKDIR}/${MODULE_DIR}/${item}.cpp"
+    OBJ_FILE="${SRS_OBJS}/${MODULE_DIR}/${item}.o"
+    MODULE_OBJS="${MODULE_OBJS[@]} ${MODULE_DIR}/${item}.cpp"
     if [ -f ${CPP_FILE} ]; then
         echo "${OBJ_FILE}: \$(${DEPS_NAME}) ${CPP_FILE} " >> ${FILE}
         echo "	\$(CXX) -c \$(CXXFLAGS) ${DEFINES}\\" >> ${FILE}
@@ -94,6 +91,6 @@ done
 echo "" >> ${FILE}
 
 # parent Makefile, to create module output dir before compile it.
-echo "	@mkdir -p ${SRS_OBJS_DIR}/${MODULE_DIR}" >> ${SRS_WORKDIR}/${SRS_MAKEFILE}
+echo "	@mkdir -p ${SRS_OBJS}/${MODULE_DIR}" >> ${SRS_MAKEFILE}
 
 echo -n "Generate modules ${MODULE_ID} ok"; echo '!';