+++ trunk/auto/utest.sh
@@ -2,26 +2,31 @@
 #
 # params:
 #     $SRS_OBJS the objs directory to store the Makefile. ie. ./objs
-#     $SRS_OBJS_DIR the objs directory for Makefile. ie. objs
-#     $SRS_MAKEFILE the makefile name. ie. Makefile
+#     $SRS_OBJS the objs directory for Makefile. ie. objs
 #
 #     $APP_NAME the app name to output. ie. srs_utest
 #     $MODULE_DIR the src dir of utest code. ie. src/utest
 #     $LINK_OPTIONS the link options for utest. ie. -lpthread -ldl
 
-FILE=${SRS_OBJS}/utest/${SRS_MAKEFILE}
+FILE=${SRS_OBJS}/${SRS_PLATFORM}/utest/Makefile
 # create dir for Makefile
-mkdir -p ${SRS_OBJS}/utest
+mkdir -p ${SRS_OBJS}/${SRS_PLATFORM}/utest
 
 # the prefix to generate the objs/utest/Makefile
 # dirs relative to current dir(objs/utest), it's trunk/objs/utest
 # trunk of srs, which contains the src dir, relative to objs/utest, it's trunk
 SRS_TRUNK_PREFIX=../../..
-# gest dir, relative to objs/utest, it's trunk/objs/gtest
-GTEST_DIR=${SRS_TRUNK_PREFIX}/${SRS_OBJS_DIR}/gtest
+# gest dir, relative to objs/utest, it's trunk/objs/{Platform}/gtest
+GTEST_DIR=../3rdpatry/gtest/googletest
+
+# Whether enable C++11 or higher versions.
+# For linux, always use C++11 for gtest required, see https://github.com/google/googletest
+# For cygwin64, ignore because it use -std=gnu++11 by default.
+SRS_CPP_VERSION="-std=c++11"
+if [[ $SRS_CYGWIN64 == YES ]]; then SRS_CPP_VERSION="-std=gnu++11"; fi
 
 cat << END > ${FILE}
-# user must run make the ${SRS_OBJS_DIR}/utest dir
+# user must run make the ${SRS_OBJS}/utest dir
 # at the same dir of Makefile.
 
 # A sample Makefile for building Google Test and using it in user
@@ -52,13 +57,13 @@ CXX = ${SRS_TOOL_CXX}
 CPPFLAGS += -I\$(GTEST_DIR)/include
 
 # Flags passed to the C++ compiler.
-CXXFLAGS += ${CXXFLAGS} ${UTEST_EXTRA_DEFINES} -Wno-unused-private-field -Wno-unused-command-line-argument
-# Always use C++11 for gtest required, see https://github.com/google/googletest
-CXXFLAGS += -std=c++11
+CXXFLAGS += ${CXXFLAGS} ${UTEST_EXTRA_DEFINES}
+CXXFLAGS += -Wno-unused-private-field -Wno-unused-command-line-argument
+CXXFLAGS += ${SRS_CPP_VERSION}
 
 # All tests produced by this Makefile.  Remember to add new tests you
 # created to the list.
-TESTS = ${SRS_TRUNK_PREFIX}/${SRS_OBJS_DIR}/${APP_NAME}
+TESTS = ${SRS_TRUNK_PREFIX}/${SRS_OBJS}/${APP_NAME}
 
 # All Google Test headers.  Usually you shouldn't change this
 # definition.
@@ -139,7 +144,7 @@ echo "# Depends, the depends objects" >> ${FILE}
 echo -n "SRS_UTEST_DEPS = " >> ${FILE}
 for item in ${MODULE_OBJS[*]}; do
     FILE_NAME=${item%.*}
-    echo -n "${SRS_TRUNK_PREFIX}/${SRS_OBJS_DIR}/${FILE_NAME}.o " >> ${FILE}
+    echo -n "${SRS_TRUNK_PREFIX}/${SRS_OBJS}/${FILE_NAME}.o " >> ${FILE}
 done
 echo "" >> ${FILE}; echo "" >> ${FILE}
 #
@@ -183,12 +188,12 @@ echo "" >> ${FILE}; echo "" >> ${FILE}
 #
 echo "# generate the utest binary" >> ${FILE}
 cat << END >> ${FILE}
-${SRS_TRUNK_PREFIX}/${SRS_OBJS_DIR}/${APP_NAME} : \$(SRS_UTEST_DEPS) ${MODULE_OBJS} gtest.a
+${SRS_TRUNK_PREFIX}/${SRS_OBJS}/${APP_NAME} : \$(SRS_UTEST_DEPS) ${MODULE_OBJS} gtest.a
 	\$(CXX) -o \$@ \$(CPPFLAGS) \$(CXXFLAGS) \$^ \$(DEPS_LIBRARIES_FILES) ${LINK_OPTIONS}
 END
 
 #####################################################################################
 # parent Makefile, to create module output dir before compile it.
-echo "	@mkdir -p ${SRS_OBJS_DIR}/utest" >> ${SRS_WORKDIR}/${SRS_MAKEFILE}
+echo "	@mkdir -p ${SRS_OBJS}/utest" >> ${SRS_MAKEFILE}
 
 echo -n "Generate utest ok"; echo '!';