+++ trunk/src/app/srs_app_rtmp_conn.cpp
@@ -557,9 +557,12 @@ srs_error_t SrsRtmpConn::stream_service_cycle()
     srs_assert(source != NULL);
 
     bool enabled_cache = _srs_config->get_gop_cache(req->vhost);
-    srs_trace("source url=%s, ip=%s, cache=%d, is_edge=%d, source_id=%s/%s",
-        req->get_stream_url().c_str(), ip.c_str(), enabled_cache, info->edge, source->source_id().c_str(), source->pre_source_id().c_str());
+    int gcmf = _srs_config->get_gop_cache_max_frames(req->vhost);
+    srs_trace("source url=%s, ip=%s, cache=%d/%d, is_edge=%d, source_id=%s/%s",
+        req->get_stream_url().c_str(), ip.c_str(), enabled_cache, gcmf, info->edge, source->source_id().c_str(),
+        source->pre_source_id().c_str());
     source->set_cache(enabled_cache);
+    source->set_gop_cache_max_frames(gcmf);
     
     switch (info->type) {
         case SrsRtmpConnPlay: {